requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cohere==5.15.0
colorama==0.4.6
fastavro==1.11.1
filelock==3.18.0
Flask==3.1.1
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.4
sniffio==1.3.1
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
